timesheet/timesheet/requirements.txt
djangorestframework
pymongo==3.12.1
django-cors-headers==3.13.0
django-mongo-auth==0.1.3
sqlparse==0.2.4
PyJWT==1.7.1
lazy==1.5
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
djangorestframework-jwt==1.11.0